Remove colors which are unreadable on white or black background
I removed `white` and `yellow` that are unreadable on white background, as well as `navy` which is hard to read on black background.

I introduced the `grey` that was previously omitted and should work on both backgrounds.

WARNING: I didn't test these changes. I merely checked the page of the `irc-colors` NPM package ( https://www.npmjs.com/package/irc-colors ) assuming that the reported colors are indeed working. Please test this PR before merging!